Move ApplicationTests into integration as ApplicationITest
Retarget the TestClassTaxonomyTest direct-annotation fixture to the new path.
